'CMakeLists.txt' (JetInterface-00-00-43)
	* JetInterface-00-00-43
	* ATLJETMET-376
	* JetInterface/IJetUpdateJvt.h: Add this interface for updating JVT.

2015-05-08 S. Schramm
	* JetInterface-00-00-42
	* IJetUncertaintiesTool: fixed pileup reference retrieval (doesn't affect uncertainty)

+++ Reconstruction/Jet/JetInterface/JetInterface/IJetUncertaintiesTool.h
@@ -28,7 +28,7 @@ class IJetUncertaintiesTool : virtual public asg::IAsgTool
         virtual std::string getPath()         const = 0;
         virtual std::string getAnalysisFile() const = 0;
         virtual float       getRefMu()        const = 0;
-        virtual unsigned    getRefNPV()       const = 0;
+        virtual float       getRefNPV()       const = 0;
 
         // Retrieve component information
         virtual size_t      getNumComponents() const = 0;
